editing type of id field in AppUser and UserPrincipal, added crypt passwords for User and Admin in DB init script

+++ src/main/java/ru/neustupov/restvotingwithspringbootandreact/model/AppUser.java
@@ -47,7 +47,7 @@ public class AppUser extends AbstractNamedEntity {
     public AppUser() {
     }
 
-    public AppUser(Integer id, String name){
+    public AppUser(Long id, String name){
         super(id, name);
     }
